upload images works

+++ src/main/java/at/htlstp/felerfrei/services/ImageLocationService.java
@@ -3,19 +3,17 @@ package at.htlstp.felerfrei.services;
 import at.htlstp.felerfrei.domain.Image;
 import at.htlstp.felerfrei.persistence.ImageRepository;
 import lombok.SneakyThrows;
-import org.assertj.core.util.Files;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service("imageLocationService")
 public class ImageLocationService implements FileService {
 
-    private static final String RESOURCE_DIR = "/";
+    private static final String RESOURCE_DIR = "images";
 
     private final ImageRepository repository;
 
@@ -35,11 +33,6 @@ public class ImageLocationService implements FileService {
 
         var path = saved.getPath();
 
-        var dir = new File(parseDirectory(path));
-        dir.mkdirs();
-        var tmp = new File(path);
-        tmp.createNewFile();
-
         var wasSavedToFilesystem = imageService.saveImage(file, path);
         if (!wasSavedToFilesystem) {
             delete(saved.getId());
@@ -48,18 +41,13 @@ public class ImageLocationService implements FileService {
         return saved.getId();
     }
 
-    private String parseDirectory(String file) {
-        return file.substring(0, file.lastIndexOf("/"));
-    }
-
-    private String praseFile(String file) {
-        return file.substring(file.lastIndexOf("/") + 1);
-    }
-
     @Override
     public void delete(int id) throws NoSuchElementException {
         var found = repository.findById(id);
-        repository.delete(found.orElseThrow(() -> new NoSuchElementException("Image not found")));
+        if(found.isEmpty()) {
+            throw new NoSuchElementException("Image not found");
+        }
+        repository.delete(found.get());
         imageService.delete(found.get().getPath());
     }
